Intentando hacer uso del websocket 🔄

+++ WebSocket-Back/src/main/java/com/example/udea/tablero/model/Coordenadas.java
@@ -1,5 +1,5 @@
 package com.example.udea.tablero.model;
 
-public record Coordenadas(int x, int y) {
+public record Coordenadas(double x, double y) {
   
 }
